backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.7
google-cloud-storage==2.10.0
python-dotenv==1.0.0
pydantic==2.11.1
pydantic-settings==2.8.1
ultralytics
opencv-python-headless
